requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cachetools==5.3.3
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
discord.py==2.4.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
gspread==6.1.2
httplib2==0.22.0
idna==3.2
multidict==5.1.0
oauthlib==3.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.6.3
